rrlfd/requirements.txt
absl-py
dm-acme[tensorflow]
dm-sonnet
gym
numpy
tensorflow
tensorflow_probability
